В отдельной таблице БД добавляется количество каждого товара в каждом заказе

+++ Application/Queries/ViewModels/OrderGetToClientDTO.cs
@@ -11,7 +11,7 @@ namespace Application.Queries.ViewModels
         public Guid Id { get; set; }
 
         public bool PaymentConfirmation { get; set; }
-        public int TotalSum { get; set; }
+        public decimal TotalSum { get; set; }
 
         public ICollection<Guid>? AutomobilesId { get; set; } 
         public ICollection<string> AutomobilesName { get; set; }
